baselines/requirements.txt
tqdm
numpy
scipy
torch==1.5.1
torchvision==0.6.1
transformers==2.8.0
pytorch_transformers==1.0.0
word2number==1.1
lmdb==0.94
tensorflow==2.5.0
tensorpack==0.9.4
easydict==1.9
PyYAML==5.4
jsonlines==1.2.0
json-lines==0.5.0
matplotlib
Cython
python-prctl
msgpack
msgpack-numpy
opencv-python==4.2.0.34
ninja
yacs
numpyencoder
